Changing Structure
- Changed the whole strucutre
- Prepared for starting of a barebone
- Added configs (eslint & prettier)
