+++ .prettierrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+   trailingComma: 'es5',
+   printWidth: 80,
+   tabWidth: 3,
+   singleQuote: true,
+   vueIndentScriptAndStyle: true,
+   arrowParens: 'always',
+   bracketSpacing: true,
+};
